requirements.txt
Django>=5.0.4,<5.1
uWSGI>=2.0.25.1,<2.1
djangorestframework>=3.15.1,<3.16
psycopg2>=2.9.9,<3
qrcode>=7.4.2,<7.5
pillow>=10.3.0,<10.4